refactor: swap utils to rs-fevm-utils (#30)
---------
