+++ cli/src/bin/saturn-contracts.rs
@@ -1,5 +1,4 @@
 use cli::commands::Cli;
-use cli::utils::banner;
 use log::{error, info};
 
 #[tokio::main]
@@ -16,4 +15,23 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     }
 
     Ok(())
+}
+
+pub fn banner() {
+    info!(
+        "{}",
+        format!(
+            "
+            _|_|_|              _|                                    
+            _|          _|_|_|  _|_|_|_|  _|    _|  _|  _|_|  _|_|_|    
+              _|_|    _|    _|    _|      _|    _|  _|_|      _|    _|  
+                  _|  _|    _|    _|      _|    _|  _|        _|    _|  
+            _|_|_|      _|_|_|      _|_|    _|_|_|  _|        _|    _|      
+
+        -----------------------------------------------------------
+        Saturn smart contracts 🪐.
+        -----------------------------------------------------------
+        "
+        )
+    );
 }
